Made all the functions within the function class static to remove errors.
Then I called upon the fucntions to make them operational. The way I did this is by replacing the outputs texts label with the text "output :", so that the users knows this is the output to their action, then following this text I called the required function, passed in the parsed paramters that I are passed on from the two text boxes.